Allow to pass additional environment variables
Allow to pass additional environment variables to the operator via the helm chart deployment. 

OPERATOR_ALL_NAMESPACES or for example OPERATOR_MODE
These are currently "internal" only and hence I do not want to add switches to the helm-chart by default.
By allowing to add arbitrary environment variables we do not expose this functionality to the outside when it's currently not official.
The helm-chart currently does not allow any env variables.